// app/shared/models/cotacao/cotacao-rodada-proposta.ts
import { CotacaoRodadaPropostaMotivoDesclassificacao } from '@shared/models/cotacao/cotacao-rodada-proposta-motivo-desclassificacao';
import { Fornecedor, UnidadeMedida } from '..';
import { Arquivo } from '../arquivo';
import { CondicaoPagamento } from '../condicao-pagamento';
import { ClassificacaoPreco } from '../enums/classificacao-preco';
import { TipoFrete } from '../enums/tipo-frete';
import { UnidadeMedidaTempo } from '../enums/unidade-medida-tempo';
import { Usuario } from '../usuario';
import { PessoaJuridica } from './../pessoa-juridica';
import { Transportadora } from './../transportadora';

export class CotacaoRodadaProposta {
  idCotacaoRodadaProposta: number;
  idCotacaoRodada: number;
  idCotacaoParticipante: number;
  idCotacaoItem: number;
  idCondicaoPagamento: number;
  condicaoPagamento: CondicaoPagamento;
  marca: string;
  ativo: boolean;
  ncm: string;
  ca: string;
  quantidadeDisponivel: number;
  modelo: string;
  precoBruto: number;
  precoLiquido: number;
  precoUnidade: number;
  valorFrete: number;
  dataEntregaDisponivel: string;
  ipiAliquota: number;
  pisAliquota: number;
  confinsAliquota: number;
  icmsAliquota: number;
  difalAliquota: number;
  stAliquota: number;
  csllAliquota: number;
  issAliquota: number;
  irAliquota: number;
  inssAliquota: number;
  garantia: number;
  unidadeMedidaGarantia: UnidadeMedidaTempo;
  observacao: string;
  anexos: Array<Arquivo>;
  vencedor: boolean;
  recomendado: boolean;
  desclassificado: boolean;
  classificacaoPreco: ClassificacaoPreco;
  fornecedor: PessoaJuridica;
  statusFornecedor: Fornecedor;
  incoterms: TipoFrete;
  enviada: boolean;
  idUsuarioResponsavelEnvioProposta: number;
  usuarioResponsavelEnvioProposta: Usuario;
  dataHoraEnvioProposta: string;
  ultimaAlteracao: string;
  idUsuarioLiberacaoReenvio: number;
  dataHoraLiberacaoReenvio: string;
  idCotacaoRodadaPropostaPai: number;
  cotacaoRodadaPropostaMotivoDesclassificacao: CotacaoRodadaPropostaMotivoDesclassificacao;
  embalagemEmbarque: number;
  idUnidadeMedidaEmbalagemEmbarque: number;
  unidadeMedidaEmbalagemEmbarque: UnidadeMedida;
  faturamentoMinimo: number;
  prazoEntrega: number;
  idTransportadora: number;
  transportadora: Transportadora;
  idIva: number;
  desconsideradoPreco: boolean;
  idPedidoGerado?: number;
  diferencaPrecoUltimaCompra?: number;
}